squelch some warnings
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@7935 8aca7d54-2c30-db11-9de9-000461428c89

+++ dataobj/tabfile.cc
@@ -344,10 +344,10 @@ bool tabfile_t::read(tabfileobj_t &objinfo)
 								}
 								combination[j]%=parameter_values[j];
 							}
-							sprintf(line_expand, "%.*s%d", param[0]-line, line, parameter_value[0][combination[0]]);
+							sprintf(line_expand, "%.*s%d", (int)(param[0]-line), line, parameter_value[0][combination[0]]);
 							for(int i=1; i<parameters; i++) {
 								char *prev_end = param[i-1]+parameter_length[i-1];
-								sprintf(buffer, "%.*s%d", param[i]-prev_end, prev_end, parameter_value[i][combination[i]]);
+								sprintf(buffer, "%.*s%d", (int)(param[i]-prev_end), prev_end, parameter_value[i][combination[i]]);
 								strcat(line_expand, buffer);
 							}
 							strcat(line_expand, param[parameters-1]+parameter_length[parameters-1]);
@@ -367,10 +367,10 @@ bool tabfile_t::read(tabfileobj_t &objinfo)
 								expansion_value[i] = calculate(buffer, parameter_value, parameters, combination);
 							}
 
-							sprintf(delim_expand, "%.*s%d", expansion[0]-delim, delim, expansion_value[0]);
+							sprintf(delim_expand, "%.*s%d", (int)(expansion[0]-delim), delim, expansion_value[0]);
 							for(int i=1; i<expansions; i++) {
 								char *prev_end = expansion[i-1]+expansion_length[i-1]+2;
-								sprintf(buffer, "%.*s%d", expansion[i]-prev_end, prev_end, expansion_value[i]);
+								sprintf(buffer, "%.*s%d", (int)(expansion[i]-prev_end), prev_end, expansion_value[i]);
 								strcat(delim_expand, buffer);
 							}
 							strcat(delim_expand, expansion[expansions-1]+expansion_length[expansions-1]+2);
@@ -568,9 +568,9 @@ void tabfile_t::add_operator_brackets(char *expression, char *processed)
 			if(expression_end==NULL) expression_end = expression_pos;
 
 			// construct expression with brackets around 'a operator b'
-			sprintf(buffer,"%.*s(%.*s%.*s)%s",	expression_start-processed+1, processed,
-								expression_ptr-expression_start-1, expression_start+1,
-								expression_end-expression_ptr, expression_ptr,
+			sprintf(buffer,"%.*s(%.*s%.*s)%s",	(int)(expression_start-processed+1), processed,
+								(int)(expression_ptr-expression_start-1), expression_start+1,
+								(int)(expression_end-expression_ptr), expression_ptr,
 								expression_end);
 
 			strcpy(processed, buffer);
